# Vectorize string.IndexOf(..., StringComparison.Ordinal)