--- a/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.Windows.cs
+++ b/src/System.Private.CoreLib/shared/System/Globalization/CompareInfo.Windows.cs
@@ -4,9 +4,9 @@
 
 using System.Buffers;
 using System.Diagnostics;
-using System.Security;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+
+using Internal.Runtime.CompilerServices;
 
 namespace System.Globalization
 {
@@ -342,12 +342,19 @@ namespace System.Globalization
 
             if ((options & CompareOptions.Ordinal) != 0)
             {
-                int retValue = FastIndexOfString(source, target, startIndex, count, target.Length, findLastIndex: false);
+                int retValue = SpanHelpers.IndexOf(
+                    ref Unsafe.Add(ref source.GetRawStringData(), startIndex),
+                    count,
+                    ref target.GetRawStringData(),
+                    target.Length);
+
                 if (retValue >= 0)
                 {
+                    retValue += startIndex;
                     if (matchLengthPtr != null)
                         *matchLengthPtr = target.Length;
                 }
+
                 return retValue;
             }
             else
@@ -388,7 +395,7 @@ namespace System.Globalization
 
             if ((options & CompareOptions.Ordinal) != 0)
             {
-                return FastIndexOfString(source, target, startIndex, count, target.Length, findLastIndex: true);
+                return FastLastIndexOfString(source, target, startIndex, count, target.Length);
             }
             else
             {
@@ -462,75 +469,43 @@ namespace System.Globalization
         private const int FIND_FROMSTART = 0x00400000;
         private const int FIND_FROMEND = 0x00800000;
 
-        // TODO: Instead of this method could we just have upstack code call IndexOfOrdinal with ignoreCase = false?
-        private static unsafe int FastIndexOfString(string source, string target, int startIndex, int sourceCount, int targetCount, bool findLastIndex)
+        // TODO: Instead of this method could we just have upstack code call LastIndexOfOrdinal with ignoreCase = false?
+        private static unsafe int FastLastIndexOfString(string source, string target, int startIndex, int sourceCount, int targetCount)
         {
             int retValue = -1;
 
-            int sourceStartIndex = findLastIndex ? startIndex - sourceCount + 1 : startIndex;
+            int sourceStartIndex = startIndex - sourceCount + 1;
 
             fixed (char* pSource = source, spTarget = target)
             {
                 char* spSubSource = pSource + sourceStartIndex;
 
-                if (findLastIndex)
+                int startPattern = (sourceCount - 1) - targetCount + 1;
+                if (startPattern < 0)
+                    return -1;
+
+                char patternChar0 = spTarget[0];
+                for (int ctrSrc = startPattern; ctrSrc >= 0; ctrSrc--)
                 {
-                    int startPattern = (sourceCount - 1) - targetCount + 1;
-                    if (startPattern < 0)
-                        return -1;
+                    if (spSubSource[ctrSrc] != patternChar0)
+                        continue;
 
-                    char patternChar0 = spTarget[0];
-                    for (int ctrSrc = startPattern; ctrSrc >= 0; ctrSrc--)
+                    int ctrPat;
+                    for (ctrPat = 1; ctrPat < targetCount; ctrPat++)
                     {
-                        if (spSubSource[ctrSrc] != patternChar0)
-                            continue;
-
-                        int ctrPat;
-                        for (ctrPat = 1; ctrPat < targetCount; ctrPat++)
-                        {
-                            if (spSubSource[ctrSrc + ctrPat] != spTarget[ctrPat])
-                                break;
-                        }
-                        if (ctrPat == targetCount)
-                        {
-                            retValue = ctrSrc;
+                        if (spSubSource[ctrSrc + ctrPat] != spTarget[ctrPat])
                             break;
-                        }
                     }
-
-                    if (retValue >= 0)
+                    if (ctrPat == targetCount)
                     {
-                        retValue += startIndex - sourceCount + 1;
+                        retValue = ctrSrc;
+                        break;
                     }
                 }
-                else
+
+                if (retValue >= 0)
                 {
-                    int endPattern = (sourceCount - 1) - targetCount + 1;
-                    if (endPattern < 0)
-                        return -1;
-
-                    char patternChar0 = spTarget[0];
-                    for (int ctrSrc = 0; ctrSrc <= endPattern; ctrSrc++)
-                    {
-                        if (spSubSource[ctrSrc] != patternChar0)
-                            continue;
-                        int ctrPat;
-                        for (ctrPat = 1; ctrPat < targetCount; ctrPat++)
-                        {
-                            if (spSubSource[ctrSrc + ctrPat] != spTarget[ctrPat])
-                                break;
-                        }
-                        if (ctrPat == targetCount)
-                        {
-                            retValue = ctrSrc;
-                            break;
-                        }
-                    }
-
-                    if (retValue >= 0)
-                    {
-                        retValue += startIndex;
-                    }
+                    retValue += startIndex - sourceCount + 1;
                 }
             }