… src/System.Private.CoreLib/shared/System/SpanHelpers.Byte.cs
… src/System.Private.CoreLib/shared/System/SpanHelpers.Char.cs
… src/System.Private.CoreLib/shared/System/String.Searching.cs