x  2.5 times faster at position 0
x 11.2 times faster at position 1024

For

```csharp
// where comparisonType == StringComparison.Ordinal 
string.IndexOf(string value, StringComparison comparisonType)
string.IndexOf(string value, int startIndex, StringComparison comparisonType)
string.IndexOf(string value, int startIndex, int count, StringComparison comparisonType)

// where T : char
IndexOf<T>(this Span<T> span, ReadOnlySpan<T> value)
IndexOf<T>(this ReadOnlySpan<T> span, ReadOnlySpan<T> value)

// where comparisonType == StringComparison.Ordinal
IndexOf(this ReadOnlySpan<char> span, ReadOnlySpan<char> value, StringComparison comparisonType)

// where options = CompareOptions.Ordinal 
CompareInfo.IndexOf(string source, string value, CompareOptions options)
CompareInfo.IndexOf(string source, string value, int startIndex, CompareOptions options)
CompareInfo.IndexOf(string source, string value, int startIndex, int count, CompareOptions options)
```

https://gist.github.com/benaadams/177d989fea110f51fab14cd708239852
```diff
           Method | Position |      Mean |
 ---------------- |--------- |----------:|
- IndexOf_Ordinal |        0 |  35.86 ns |
+ IndexOf_Ordinal |        0 |  14.15 ns |
- IndexOf_Ordinal |        7 |  42.69 ns |
+ IndexOf_Ordinal |        7 |  19.00 ns |
- IndexOf_Ordinal |        8 |  43.13 ns |
+ IndexOf_Ordinal |        8 |  18.90 ns |
- IndexOf_Ordinal |       15 |  49.67 ns |
+ IndexOf_Ordinal |       15 |  19.61 ns |
- IndexOf_Ordinal |       16 |  47.58 ns |
+ IndexOf_Ordinal |       16 |  19.64 ns |
- IndexOf_Ordinal |       31 |  60.92 ns |
+ IndexOf_Ordinal |       31 |  21.09 ns |
- IndexOf_Ordinal |       32 |  62.66 ns |
+ IndexOf_Ordinal |       32 |  21.11 ns |
- IndexOf_Ordinal |       63 | 100.10 ns |
+ IndexOf_Ordinal |       63 |  22.45 ns |
- IndexOf_Ordinal |       64 | 100.82 ns |
+ IndexOf_Ordinal |       64 |  22.54 ns |
- IndexOf_Ordinal |      127 | 156.94 ns |
+ IndexOf_Ordinal |      127 |  27.10 ns |
- IndexOf_Ordinal |      128 | 155.85 ns |
+ IndexOf_Ordinal |      128 |  26.07 ns |
- IndexOf_Ordinal |      255 | 265.79 ns |
+ IndexOf_Ordinal |      255 |  32.93 ns |
- IndexOf_Ordinal |      256 | 266.91 ns |
+ IndexOf_Ordinal |      256 |  32.98 ns |
- IndexOf_Ordinal |     1023 | 918.89 ns |
+ IndexOf_Ordinal |     1023 |  81.13 ns |
- IndexOf_Ordinal |     1024 | 916.90 ns |
+ IndexOf_Ordinal |     1024 |  81.20 ns |
```

Contributes to https://github.com/dotnet/coreclr/issues/19672

/cc @jkotas, @krwq, @tarekgh