@benaadams 2018-11-19
OSX10.12 infra issue: https://github.com/xunit/xunit/issues/1855