@benaadams 2018-11-19
OSX Fail
```
.GetFileSizeEx: ERROR -> SetEndOfFile failed due to lack of disk space
```