@benaadams 2018-11-19
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test please