@krwq 2018-11-19
I assume there already are test cases checking for all of the edge cases we could hit here?