@MattGal 2018-11-19
@benaadams @tarekgh : note that running out of disk space on an OSX machine does not make OSX CI generally broken since after finishing their current Jenkins agent job these machines will in most cases (except when jusssssst enough cleanup happens after the agent finishes) will not take more work.  In the retry case, the machine its running on currently still has 18 GB free space and will hopefully succeed.  

Jenkins in general rarely cleans up workspaces, so it's a constant battle to clean up these machines.